requirements.txt
pandas~=2.2.3
python-dotenv~=0.21.0
sqlalchemy~=2.0.37
numpy~=1.26.4
selenium~=4.27.1
beautifulsoup4~=4.12.2